Parses filter, select, rename and reshape call

Parse filter, select, rename and reshape call

get_dplyr_call(
  selector_list,
  idx = 1L,
  join_keys = list(),
  dplyr_call_data = get_dplyr_call_data(selector_list, join_keys = join_keys),
  datasets = NULL
)

Arguments

selector_list

(reactive)
output from data_extract_multiple_srv() or a reactive named list of outputs from data_extract_srv(). When using a reactive named list, the names must be identical to the shiny ids of the respective data_extract_ui().

idx

optional (integer) current selector index in all selectors list

join_keys

(named list) nested list of keys used for joining

dplyr_call_data

(list) simplified selectors with aggregated set of filters, selections, reshapes etc. All necessary data for merging

datasets

(NULL or FilteredData).

Value

(call) filter, select, rename and reshape call

Examples

# one dataset
teal.transform:::get_dplyr_call(
  list(list(
    dataname = "ADSL",
    filters = NULL,
    select = character(0),
    keys = teal.data::get_cdisc_keys("ADSL"),
    reshape = FALSE,
    internal_id = "test1"
  ))
)
#> ADSL_FILTERED
teal.transform:::get_dplyr_call(
  list(list(
    dataname = "ADSL",
    filters = list(list(columns = "SEX", selected = list("F", "M"))),
    select = character(0),
    keys = teal.data::get_cdisc_keys("ADSL"),
    reshape = FALSE,
    internal_id = "test1"
  ))
)
#> ADSL_FILTERED %>% dplyr::filter(SEX %in% c("F", "M"))
teal.transform:::get_dplyr_call(
  list(list(
    dataname = "ADSL",
    filters = list(list(columns = "SEX", selected = list("F", "M"))),
    select = c("AVAL"),
    keys = teal.data::get_cdisc_keys("ADSL"),
    reshape = FALSE,
    internal_id = "test1"
  ))
)
#> ADSL_FILTERED %>% dplyr::filter(SEX %in% c("F", "M")) %>% dplyr::select(AVAL)

# two datasets with rename part
teal.transform:::get_dplyr_call(
  list(
    list(
      dataname = "ADSL",
      filters = NULL,
      select = c("COL_1", "COL_2"),
      keys = teal.data::get_cdisc_keys("ADSL"),
      reshape = FALSE,
      internal_id = "test1"
    ),
    list(
      dataname = "ADSL",
      filters = NULL,
      select = c("COL_2", "COL_3"),
      keys = teal.data::get_cdisc_keys("ADSL"),
      reshape = FALSE,
      internal_id = "test2"
    )
  ),
  idx = 1L
)
#> ADSL_FILTERED %>% dplyr::select(COL_1, COL_2) %>% dplyr::rename(test1.COL_2 = COL_2)

# long dataset with reshape part
teal.transform:::get_dplyr_call(
  list(list(
    dataname = "ADLB",
    filters = list(list(
      columns = c("PARAMCD", "AVISIT"),
      selected = list(c("ALBCV", "SCREENING"), c("ALBCV", "BASELINE"))
    )),
    select = c("AVAL"),
    keys = teal.data::get_cdisc_keys("ADLB"),
    reshape = TRUE,
    internal_id = "test1"
  ))
)
#> ADLB_FILTERED %>% dplyr::filter((PARAMCD == "ALBCV" & AVISIT == 
#>     "SCREENING") | (PARAMCD == "ALBCV" & AVISIT == "BASELINE")) %>% 
#>     dplyr::select(PARAMCD, AVISIT, AVAL) %>% tidyr::pivot_longer(cols = "AVAL", 
#>     names_to = "MEASURE", values_to = "VALUE") %>% tidyr::unite(KEY, 
#>     MEASURE, PARAMCD, AVISIT) %>% tidyr::pivot_wider(names_from = "KEY", 
#>     values_from = "VALUE")