Returns the initial values for the vals widget of a filter_spec object.

Returns the initial values for the vals widget of a filter_spec object.

get_initial_filter_values(filter, datasets)

Arguments

filter

(filter_spec) the filter generated by a call to filter_spec()

datasets

(FilteredData) the datasets

Value

list of two values: col and vals

Examples

filtered_data <- teal.slice:::FilteredData$new()$set_dataset(
  teal.data::dataset("iris", utils::head(iris))
)
filter <- filter_spec(vars = colnames(iris)[1])
filter$dataname <- "iris"
shiny::isolate(
  teal.transform:::get_initial_filter_values(filter = filter, datasets = filtered_data)
)
#> $choices
#> number of choices: 6 
#> 
#> 5.1: 
#> 4.9: 
#> 4.7: 
#> 4.6: 
#> 5: 
#> 5.4: 
#> 
#>